fairscan/backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
scipy
shap
aif360
reportlab
google-generativeai
python-multipart
python-dotenv
numpy
